requirements.txt
pillow
numpy
tifffile